move fs-extra mock to dedicated file
