+++ lib/v4/migration-helpers/__mocks__/mock-package-json-v3.json
