+++ lib/v4/migration-helpers/__tests__/convert-models-to-content-types.test.js
@@ -1,11 +1,6 @@
-jest.mock('fs-extra', () => ({
-  exists: jest.fn(() => Promise.resolve()),
-  readdir: jest.fn(() => Promise.resolve()),
-  readJSON: jest.fn(() => Promise.resolve()),
-  remove: jest.fn(() => Promise.resolve()),
-  ensureFile: jest.fn(() => Promise.resolve()),
-  writeJSON: jest.fn(() => Promise.resolve()),
-}));
+jest.mock('fs-extra', () => {
+  return require('../__mocks__/mock-fs.js');
+});
 
 const { join } = require('path');
 const fs = require('fs-extra');
