+++ lib/v4/migration-helpers/__tests__/update-api-policies.test.js
@@ -1,8 +1,6 @@
-jest.mock('fs-extra', () => ({
-  exists: jest.fn(() => Promise.resolve()),
-  readdir: jest.fn(() => Promise.resolve()),
-  copy: jest.fn(() => Promise.resolve()),
-}));
+jest.mock('fs-extra', () => {
+  return require('../__mocks__/mock-fs')
+});
 
 const { join } = require('path');
 const fs = require('fs-extra');
