+++ lib/v4/migration-helpers/__tests__/update-package-dependencies.test.js
@@ -2,9 +2,9 @@ jest.mock('axios', () => ({
   get: jest.fn(() => Promise.resolve()),
 }));
 
-jest.mock('fs-extra', () => ({
-  writeJSON: jest.fn(() => Promise.resolve()),
-}));
+jest.mock('fs-extra', () => {
+  return require('../__mocks__/mock-fs')
+});
 
 const path = require('path');
 const fs = require('fs-extra');
@@ -13,7 +13,7 @@ const axios = require('axios');
 const updatePackageDependencies = require('../update-package-dependencies');
 
 const packageJsonPath = path.resolve(
-  path.join(__dirname, '..', 'mocks', 'mock-package-json-v3.json')
+  path.join(__dirname, '..', '__mocks__', 'mock-package-json-v3.json')
 );
 
 describe('update api folder structure', () => {
