+++ lib/v4/migration-helpers/__tests__/update-plugin-folder-structure/create-server-index.test.js
@@ -1,7 +1,6 @@
-jest.mock('fs-extra', () => ({
-  copy: jest.fn(() => Promise.resolve()),
-  readdir: jest.fn(() => Promise.resolve()),
-}));
+jest.mock('fs-extra', () => {
+  return require('../../__mocks__/mock-fs')
+});
 
 jest.mock('../../update-plugin-folder-structure/transforms', () => ({
   importFilesToIndex: jest.fn(),
