+++ lib/v4/migration-helpers/__tests__/update-plugin-folder-structure/index.test.js
@@ -1,10 +1,6 @@
-jest.mock('fs-extra', () => ({
-  pathExists: jest.fn(() => Promise.resolve()),
-  readdir: jest.fn(() => Promise.resolve()),
-  readFile: jest.fn(() => Promise.resolve()),
-  copy: jest.fn(() => Promise.resolve()),
-  move: jest.fn(() => Promise.resolve()),
-}));
+jest.mock('fs-extra', () => {
+  return require('../../__mocks__/mock-fs')
+});
 
 jest.mock('../../update-plugin-folder-structure/utils', () => ({
   createDirectoryIndex: jest.fn(() => Promise.resolve()),
@@ -75,7 +71,7 @@ describe('update plugin folder structure', () => {
     async (directory) => {
       fs.pathExists.mockReturnValueOnce(false).mockReturnValue(true);
       const dirPath = resolve('./test-dir');
-      
+
       await updatePluginFolderStructure(dirPath);
 
       switch (directory) {
