+++ lib/v4/migration-helpers/__tests__/update-plugin-folder-structure/move-bootstrap-function.test.js
@@ -1,8 +1,6 @@
-jest.mock('fs-extra', () => ({
-  pathExists: jest.fn(() => Promise.resolve()),
-  readdir: jest.fn(() => Promise.resolve()),
-  remove: jest.fn(() => Promise.resolve()),
-}));
+jest.mock('fs-extra', () => {
+  return require('../../__mocks__/mock-fs')
+});
 
 jest.mock('../../update-plugin-folder-structure/utils/move-to-server', () => jest.fn());
