+++ lib/v4/migration-helpers/__tests__/update-plugin-folder-structure/move-to-server.test.js
@@ -1,7 +1,6 @@
-jest.mock('fs-extra', () => ({
-  move: jest.fn(() => Promise.resolve()),
-  pathExists: jest.fn(() => Promise.resolve()),
-}));
+jest.mock('fs-extra', () => {
+  return require('../../__mocks__/mock-fs')
+});
 
 const { join } = require('path');
 const fs = require('fs-extra');
