+++ lib/v4/migration-helpers/__tests__/update-routes.test.js
@@ -9,8 +9,8 @@ const { inspect } = require('util');
 const fs = require('fs-extra');
 
 const updateRoutes = require('../update-routes');
-const mockRoutesV3 = require('../mocks/mock-routes-v3');
-const mockRoutesV4 = require('../mocks/mock-routes-v4');
+const mockRoutesV3 = require('../__mocks__/mock-routes-v3');
+const mockRoutesV4 = require('../__mocks__/mock-routes-v4');
 
 describe('migrate routes to v4', () => {
   beforeEach(() => {
